+++ st2api/st2api/cmd/api.py
@@ -13,9 +13,6 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-import os
-import sys
-
 # NOTE: It's important that we perform monkey patch as early as possible before any other modules
 # are important, otherwise SSL support for MongoDB won't work.
 # See https://github.com/StackStorm/st2/issues/4832 and https://github.com/gevent/gevent/issues/1016
@@ -24,15 +21,11 @@ from st2common.util.monkey_patch import monkey_patch
 
 monkey_patch()
 
-import eventlet
 from oslo_config import cfg
-from eventlet import wsgi
 
 from st2common import log as logging
-from st2common.service_setup import setup as common_setup
-from st2common.service_setup import teardown as common_teardown
-from st2common.service_setup import deregister_service
 from st2api import config
+from st2common.openapi import api
 
 config.register_opts(ignore_errors=True)
 
@@ -43,87 +36,27 @@ from st2api.validation import validate_rbac_is_correctly_configured
 __all__ = ["main"]
 
 LOG = logging.getLogger(__name__)
-API = "api"
-
-# How much time to give to the request in progress to finish in seconds before killing them
-WSGI_SERVER_REQUEST_SHUTDOWN_TIME = 2
-
-
-def _setup():
-    capabilities = {
-        "name": "api",
-        "listen_host": cfg.CONF.api.host,
-        "listen_port": cfg.CONF.api.port,
-        "listen_ssl": cfg.CONF.api.use_ssl,
-        "type": "active",
-    }
-
-    common_setup(
-        service=API,
-        config=config,
-        setup_db=True,
-        register_mq_exchanges=True,
-        register_signal_handlers=True,
-        register_internal_trigger_types=True,
-        service_registry=True,
-        capabilities=capabilities,
-    )
-
-    # Additional pre-run time checks
-    validate_auth_cookie_is_correctly_configured()
-    validate_rbac_is_correctly_configured()
-
-
-def _run_server():
-    host = cfg.CONF.api.host
-    port = cfg.CONF.api.port
-    use_ssl = cfg.CONF.api.use_ssl
-
-    cert_file_path = os.path.realpath(cfg.CONF.api.cert)
-    key_file_path = os.path.realpath(cfg.CONF.api.key)
-
-    if use_ssl and not os.path.isfile(cert_file_path):
-        raise ValueError('Certificate file "%s" doesn\'t exist' % (cert_file_path))
-
-    if use_ssl and not os.path.isfile(key_file_path):
-        raise ValueError('Private key file "%s" doesn\'t exist' % (key_file_path))
-
-    LOG.info(
-        "(PID=%s) ST2 API is serving on %s://%s:%s.",
-        os.getpid(),
-        "https" if use_ssl else "http",
-        host,
-        port,
-    )
-
-    max_pool_size = eventlet.wsgi.DEFAULT_MAX_SIMULTANEOUS_REQUESTS
-    worker_pool = eventlet.GreenPool(max_pool_size)
-    sock = eventlet.listen((host, port))
-
-    if use_ssl:
-        sock = eventlet.wrap_ssl(
-            sock, certfile=cert_file_path, keyfile=key_file_path, server_side=True
-        )
-
-    wsgi.server(
-        sock, app.setup_app(), custom_pool=worker_pool, log=LOG, log_output=False
-    )
-    return 0
-
-
-def _teardown():
-    common_teardown()
 
 
 def main():
-    try:
-        _setup()
-        return _run_server()
-    except SystemExit as exit_code:
-        deregister_service(API)
-        sys.exit(exit_code)
-    except Exception:
-        LOG.exception("(PID=%s) ST2 API quit due to exception.", os.getpid())
-        return 1
-    finally:
-        _teardown()
+    common_setup = {
+        "register_mq_exchanges": True,
+        "register_internal_trigger_types": True,
+    }
+
+    pre_run_checks = [
+        validate_auth_cookie_is_correctly_configured,
+        validate_rbac_is_correctly_configured,
+    ]
+
+    api.run(
+        service_name="api",
+        app_config=config,
+        cfg=cfg.CONF.api,
+        app=app,
+        log=LOG,
+        use_custom_pool=False,
+        log_output=False,
+        common_setup_kwargs=common_setup,
+        pre_run_checks=pre_run_checks,
+    )