+++ st2stream/st2stream/app.py
@@ -24,70 +24,24 @@ Note: This app doesn't need access to MongoDB, just RabbitMQ.
 
 from oslo_config import cfg
 
+from st2common.openapi import app
 from st2stream import config as st2stream_config
-from st2common import log as logging
-from st2common.middleware.streaming import StreamingMiddleware
-from st2common.middleware.error_handling import ErrorHandlingMiddleware
-from st2common.middleware.cors import CorsMiddleware
-from st2common.middleware.request_id import RequestIDMiddleware
-from st2common.middleware.logging import LoggingMiddleware
-from st2common.middleware.instrumentation import RequestInstrumentationMiddleware
-from st2common.middleware.instrumentation import ResponseInstrumentationMiddleware
-from st2common.router import Router
-from st2common.constants.system import VERSION_STRING
-from st2common.service_setup import setup as common_setup
-from st2common.util import spec_loader
-
-LOG = logging.getLogger(__name__)
 
 
 def setup_app(config={}):
-    LOG.info("Creating st2stream: %s as OpenAPI app.", VERSION_STRING)
+    common_setup = {
+        "register_mq_exchanges": True,
+        "register_internal_trigger_types": False,
+        "run_migrations": False,
+    }
 
-    is_gunicorn = config.get("is_gunicorn", False)
-    if is_gunicorn:
-
-        st2stream_config.register_opts(ignore_errors=True)
-        capabilities = {
-            "name": "stream",
-            "listen_host": cfg.CONF.stream.host,
-            "listen_port": cfg.CONF.stream.port,
-            "listen_ssl": cfg.CONF.stream.use_ssl,
-            "type": "active",
-        }
-        # This should be called in gunicorn case because we only want
-        # workers to connect to db, rabbbitmq etc. In standalone HTTP
-        # server case, this setup would have already occurred.
-        common_setup(
-            service="stream",
-            config=st2stream_config,
-            setup_db=True,
-            register_mq_exchanges=True,
-            register_signal_handlers=True,
-            register_internal_trigger_types=False,
-            run_migrations=False,
-            service_registry=True,
-            capabilities=capabilities,
-            config_args=config.get("config_args", None),
-        )
-
-    router = Router(
-        debug=cfg.CONF.stream.debug, auth=cfg.CONF.auth.enable, is_gunicorn=is_gunicorn
-    )
-
-    spec = spec_loader.load_spec("st2common", "openapi.yaml.j2")
     transforms = {"^/stream/v1/": ["/", "/v1/"]}
-    router.add_spec(spec, transforms=transforms)
 
-    app = router.as_wsgi
-
-    # Order is important. Check middleware for detailed explanation.
-    app = StreamingMiddleware(app)
-    app = ErrorHandlingMiddleware(app)
-    app = CorsMiddleware(app)
-    app = LoggingMiddleware(app, router)
-    app = ResponseInstrumentationMiddleware(app, router, service_name="stream")
-    app = RequestIDMiddleware(app)
-    app = RequestInstrumentationMiddleware(app, router, service_name="stream")
-
-    return app
+    return app.setup_app(
+        service_name="stream",
+        app_config=st2stream_config,
+        oslo_cfg=cfg.CONF.stream,
+        transforms=transforms,
+        common_setup_kwargs=common_setup,
+        config=config,
+    )